Refactor Wrap Ether
